+++ tests/test_list_parameters.py
@@ -11,7 +11,7 @@ from uqtestfuns.core.registry import get_registry
 def test_default_call():
     """Test function call without any arguments."""
     list_names = list_functions(tabulate=False)
-    for name in list_names:  # type: ignore
+    for name in list_names:
         assert_call(list_parameters, name)
 
 
@@ -19,7 +19,7 @@ def test_create_function():
     """Test creating a function with parameter IDs supplied by the function."""
     list_names = list_functions(tabulate=False)
     reg = get_registry()
-    for name in list_names:  # type: ignore
+    for name in list_names:
         param_ids = list_parameters(name, tabulate=False)
         for param_id in param_ids:
             if reg[name].input_dimension is None:
@@ -31,7 +31,7 @@ def test_create_function():
 def test_no_parameter():
     """Test listing parameters when no parameter is available."""
     function_names = list_functions(parameterized=False, tabulate=False)
-    for function_name in function_names:  # type: ignore
+    for function_name in function_names:
         param_ids = list_parameters(function_name, tabulate=False)
 
         # Assertion
@@ -41,7 +41,7 @@ def test_no_parameter():
 def test_with_parameter():
     """Test listing parameters when one or more parameter is available."""
     function_names = list_functions(parameterized=True, tabulate=False)
-    for function_name in function_names:  # type: ignore
+    for function_name in function_names:
         param_ids = list_parameters(function_name, tabulate=False)
 
         # Assertion